Fix the order of output functions calls
I was wrong in issue #5 where I suggested that  `format_text()` should be called first. As can be seen from a helpful developer debugging message, the correct order it to rewrite URLs first and then format the text.

This patch fixes the order of file_rewrite_pluginfile_urls() and format_text().